* 'master' of https://github.com/AY2122S1-CS2103-T14-4/tp: (30 commits)
  Fix bug where statuses being changed out of was not outdated
  Fix tags wrapping and only allow up to 50 characters for tag names
  Change type in ui to Relationship
  Checkstyle commit
  resolves invalid date input exception #135
  Refactored Tabs to use Tab enums
  Fix checkstyle error
  [Broken at the moment] Change to enums
  Add list context switching
  Resovles #139 in UG potential enhancements
  Fixed Schedule error message #142
  Added explaination to resolve #143 in UG
  Bug resovled #145
  bug resolved #158
  Fix indentation errors
  Add missing error handling and select command error
  More checkstyle fixes
  Checkstyle fixes
  Allow >MAX_INT values for policy id
  Bug fix for client equality check
  ...

+++ src/main/java/seedu/address/model/person/Policy.java
@@ -21,7 +21,7 @@ public class Policy {
 
     public final Name name;
     public final Name insurer;
-    public final int number;
+    public final long number;
     public final Double commission;
 
     /**
@@ -32,12 +32,12 @@ public class Policy {
      * @param name A valid policy name.
      * @param commission A valid commission.
      */
-    public Policy(Name insurer, int number, Name name, double commission) {
+    public Policy(Name insurer, long number, Name name, double commission) {
         requireAllNonNull(insurer, number, name, commission);
         this.name = name;
         this.insurer = insurer;
         this.commission = Double.valueOf(commission);
-        this.number = Integer.valueOf(number);
+        this.number = Long.valueOf(number);
     }
 
     public Name getName() {
@@ -48,7 +48,7 @@ public class Policy {
         return insurer;
     }
 
-    public int getNumber() {
+    public long getNumber() {
         return number;
     }
